HV-822 split up of message check into user defined custom message and default annotation message checks
