+++ annotation-processor/src/main/java/org/hibernate/validator/ap/checks/annotationparameters/AnnotationMessageCheck.java
@@ -6,38 +6,27 @@
  */
 package org.hibernate.validator.ap.checks.annotationparameters;
 
-import java.util.Collections;
-import java.util.Set;
 import java.util.regex.Pattern;
 import javax.lang.model.element.AnnotationMirror;
-import javax.lang.model.element.AnnotationValue;
-import javax.lang.model.element.Element;
-import javax.lang.model.element.ElementKind;
-import javax.lang.model.element.ExecutableElement;
-import javax.lang.model.element.TypeElement;
-import javax.lang.model.util.Elements;
 
-import org.hibernate.validator.ap.checks.ConstraintCheckIssue;
 import org.hibernate.validator.ap.util.AnnotationApiHelper;
-import org.hibernate.validator.ap.util.CollectionHelper;
 
 /**
- * Checks that the message provided as annotation parameter is valid and gives a warning otherwise.
+ * A base class to check that the message provided as annotation parameter is valid and gives a warning otherwise.
+ * Two known implementations are {@link AnnotationUserMessageCheck} - checks that message defined by a user is valid,
+ * {@link AnnotationDefaultMessageCheck} - checks that default message is valid.
  *
  * @author Marko Bekhta
  */
-public class AnnotationMessageCheck extends AnnotationParametersAbstractCheck {
+public abstract class AnnotationMessageCheck extends AnnotationParametersAbstractCheck {
 
-	private static final String WORDS_SEPARATED_WITH_DOTS = "(\\w)+(\\.(\\w)+)*" ;
+	private static final String WORDS_SEPARATED_WITH_DOTS = "(\\w)+(\\.(\\w)+)*";
 
 	// for dots and no {} around, or one of the {} is missing
 	private static final Pattern MESSAGE_PATTERN = Pattern.compile( WORDS_SEPARATED_WITH_DOTS + "|\\{" + WORDS_SEPARATED_WITH_DOTS + "|" + WORDS_SEPARATED_WITH_DOTS + "\\}" );
 
-	private final Elements elementUtils;
-
-	public AnnotationMessageCheck(AnnotationApiHelper annotationApiHelper, Elements elementUtils) {
+	public AnnotationMessageCheck(AnnotationApiHelper annotationApiHelper) {
 		super( annotationApiHelper );
-		this.elementUtils = elementUtils;
 	}
 
 	@Override
@@ -45,35 +34,14 @@ public class AnnotationMessageCheck extends AnnotationParametersAbstractCheck {
 		return true;
 	}
 
-	@Override
-	protected Set<ConstraintCheckIssue> doCheck(Element element, AnnotationMirror annotation) {
-
-		//check if default message on annotation is correct or not:
-		if ( ElementKind.ANNOTATION_TYPE.equals( element.getKind() ) ) {
-			for ( Element innerElement : elementUtils.getAllMembers( (TypeElement) element ) ) {
-				if ( ElementKind.METHOD.equals( innerElement.getKind() ) && "message".equals( innerElement.getSimpleName().toString() ) ) {
-					if ( MESSAGE_PATTERN.matcher( ( (ExecutableElement) innerElement ).getDefaultValue().getValue().toString() ).matches() ) {
-						return CollectionHelper.asSet(
-								ConstraintCheckIssue.warning(
-										innerElement, annotation, "INVALID_MESSAGE_VALUE_ANNOTATION_PARAMETERS"
-								)
-						);
-					}
-				}
-			}
-		}
-
-		// check if the redefined by user message is correct or not:
-		AnnotationValue value = annotationApiHelper.getAnnotationValue( annotation, "message" );
-
-		if ( value != null && MESSAGE_PATTERN.matcher( (String) value.getValue() ).matches() ) {
-			return CollectionHelper.asSet(
-					ConstraintCheckIssue.warning(
-							element, annotation, "INVALID_MESSAGE_VALUE_ANNOTATION_PARAMETERS"
-					)
-			);
-		}
-
-		return Collections.emptySet();
+	/**
+	 * Verifies that message passed as parameter is valid (passes a regexp check).
+	 *
+	 * @param message a message to verify
+	 *
+	 * @return {@code true} if message is valid, {@code false} otherwise
+	 */
+	protected boolean checkMessage(String message) {
+		return MESSAGE_PATTERN.matcher( message ).matches();
 	}
 }
